+++ cli/src/config.rs
@@ -857,6 +857,9 @@ pub struct _Validator {
     // Warp the ledger to WARP_SLOT after starting the validator.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub warp_slot: Option<String>,
+    // Load a geyser config file
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub geyser_plugin_config: Option<String>,
 }
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -886,6 +889,8 @@ pub struct Validator {
     pub slots_per_epoch: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub warp_slot: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub geyser_plugin_config: Option<String>,
 }
 
 impl From<_Validator> for Validator {
@@ -911,6 +916,7 @@ impl From<_Validator> for Validator {
                 .unwrap_or(solana_sdk::rpc_port::DEFAULT_RPC_PORT),
             slots_per_epoch: _validator.slots_per_epoch,
             warp_slot: _validator.warp_slot,
+            geyser_plugin_config: _validator.geyser_plugin_config
         }
     }
 }
@@ -932,6 +938,7 @@ impl From<Validator> for _Validator {
             rpc_port: Some(validator.rpc_port),
             slots_per_epoch: validator.slots_per_epoch,
             warp_slot: validator.warp_slot,
+            geyser_plugin_config: validator.geyser_plugin_config
         }
     }
 }
@@ -999,6 +1006,7 @@ impl Merge for _Validator {
                 .slots_per_epoch
                 .or_else(|| self.slots_per_epoch.take()),
             warp_slot: other.warp_slot.or_else(|| self.warp_slot.take()),
+            geyser_plugin_config: other.geyser_plugin_config.or_else(|| self.geyser_plugin_config.take())
         };
     }
 }